+++ packages/chakra-ui/src/CheckboxWidget/CheckboxWidget.tsx
@@ -1,15 +1,8 @@
 import type { FocusEvent } from 'react';
 import type { CheckboxCheckedChangeDetails } from '@chakra-ui/react';
-import { Text } from '@chakra-ui/react';
+import { Field as ChakraField, Text } from '@chakra-ui/react';
 import type { WidgetProps, StrictRJSFSchema, RJSFSchema, FormContextType } from '@rjsf/utils';
-import {
-  ariaDescribedByIds,
-  descriptionId,
-  getTemplate,
-  labelValue,
-  schemaRequiresTrueValue,
-  getUiOptions,
-} from '@rjsf/utils';
+import { ariaDescribedByIds, descriptionId, getTemplate, getUiOptions, schemaRequiresTrueValue } from '@rjsf/utils';
 
 import { Checkbox } from '../components/ui/checkbox';
 import { Field } from '../components/ui/field';
@@ -35,11 +28,8 @@ export default function CheckboxWidget<
     options,
     uiSchema,
     schema,
+    required,
   } = props;
-  // Because an unchecked checkbox will cause html5 validation to fail, only add
-  // the "required" attribute if the field value must be "true", due to the
-  // "const" or "enum" keywords
-  const required = schemaRequiresTrueValue<S>(schema);
   const DescriptionFieldTemplate = getTemplate<'DescriptionFieldTemplate', T, S, F>(
     'DescriptionFieldTemplate',
     registry,
@@ -48,6 +38,7 @@ export default function CheckboxWidget<
   const uiOptions = getUiOptions(uiSchema);
   const isCheckbox = uiOptions.widget === 'checkbox';
   const description = isCheckbox ? undefined : (options.description ?? schema.description);
+  const inputRequired = schemaRequiresTrueValue(schema) && required;
 
   const handleChange = ({ checked }: CheckboxCheckedChangeDetails) => onChange(checked);
   const handleBlur = ({ target }: FocusEvent<HTMLInputElement | any>) => onBlur(id, target?.checked);
@@ -76,7 +67,12 @@ export default function CheckboxWidget<
         onFocus={handleFocus}
         aria-describedby={ariaDescribedByIds(id)}
       >
-        {labelValue(<Text>{label}</Text>, hideLabel || !label)}
+        {!hideLabel && label ? (
+          <Text>
+            {label}
+            {inputRequired && <ChakraField.RequiredIndicator />}
+          </Text>
+        ) : undefined}
       </Checkbox>
     </Field>
   );