+++ packages/mui/src/CheckboxWidget/CheckboxWidget.tsx
@@ -44,6 +44,7 @@ export default function CheckboxWidget<
     registry,
     options,
     uiSchema,
+    required,
   } = props;
   const DescriptionFieldTemplate = getTemplate<'DescriptionFieldTemplate', T, S, F>(
     'DescriptionFieldTemplate',
@@ -53,7 +54,7 @@ export default function CheckboxWidget<
   // Because an unchecked checkbox will cause html5 validation to fail, only add
   // the "required" attribute if the field value must be "true", due to the
   // "const" or "enum" keywords
-  const required = schemaRequiresTrueValue<S>(schema);
+  const inputRequired = schemaRequiresTrueValue<S>(schema) && required;
 
   const handleChange = (_: any, checked: boolean) => onChange(checked);
   const handleBlur: React.FocusEventHandler<HTMLButtonElement> = () => onBlur(id, value);
@@ -76,12 +77,13 @@ export default function CheckboxWidget<
       <FormControlLabel
         {...otherMuiProps}
         {...muiSlotProps?.formControlLabel}
+        required={inputRequired}
         control={
           <Checkbox
             id={id}
             name={htmlName || id}
             checked={typeof value === 'undefined' ? false : Boolean(value)}
-            required={required}
+            required={inputRequired}
             disabled={disabled || readonly}
             autoFocus={autofocus}
             onChange={handleChange}